+++ src/ImageSharp/ColorSpaces/Conversion/ColorSpaceConverter.CieLab.cs
@@ -37,7 +37,7 @@ public partial class ColorSpaceConverter
         ref CieLch sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref CieLch sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -70,7 +70,7 @@ public partial class ColorSpaceConverter
         ref CieLchuv sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref CieLchuv sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -103,7 +103,7 @@ public partial class ColorSpaceConverter
         ref CieLuv sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref CieLuv sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -136,7 +136,7 @@ public partial class ColorSpaceConverter
         ref CieXyy sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref CieXyy sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -169,7 +169,7 @@ public partial class ColorSpaceConverter
         ref CieXyz sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref CieXyz sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -201,7 +201,7 @@ public partial class ColorSpaceConverter
         ref Cmyk sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref Cmyk sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -234,7 +234,7 @@ public partial class ColorSpaceConverter
         ref Hsl sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref Hsl sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -266,7 +266,7 @@ public partial class ColorSpaceConverter
         ref Hsv sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref Hsv sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -299,7 +299,7 @@ public partial class ColorSpaceConverter
         ref HunterLab sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref HunterLab sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -332,7 +332,7 @@ public partial class ColorSpaceConverter
         ref Lms sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref Lms sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -365,7 +365,7 @@ public partial class ColorSpaceConverter
         ref LinearRgb sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref LinearRgb sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -398,7 +398,7 @@ public partial class ColorSpaceConverter
         ref Rgb sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref Rgb sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
@@ -431,7 +431,7 @@ public partial class ColorSpaceConverter
         ref YCbCr sourceRef = ref MemoryMarshal.GetReference(source);
         ref CieLab destRef = ref MemoryMarshal.GetReference(destination);
 
-        for (int i = 0; i < count; i++)
+        for (nint i = 0; i < (uint)count; i++)
         {
             ref YCbCr sp = ref Unsafe.Add(ref sourceRef, i);
             ref CieLab dp = ref Unsafe.Add(ref destRef, i);
